+++ lib/src/commands/types/mp_command_type.dart
@@ -18,6 +18,7 @@ enum MPCommandType {
   editPointType,
   moveArea,
   moveBezierLineSegment,
+  moveImageInsertConfig,
   moveLine,
   movePoint,
   moveStraightLineSegment,