+++ lib/src/widgets/mp_images_widget.dart
@@ -4,6 +4,7 @@ import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter/material.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/elements/th_element.dart';
+import 'package:mapiah/src/widgets/mp_image_operation_overlay_widget.dart';
 import 'package:mapiah/src/widgets/mp_raster_image_widget.dart';
 import 'package:mapiah/src/widgets/mp_xvi_image_widget.dart';
 
@@ -46,6 +47,29 @@ class MPImagesWidget extends StatelessWidget {
                 key: ValueKey('raster_image_${image.mpID}'),
                 th2FileEditController: th2FileEditController,
                 image: rasterImage,
+              ),
+            );
+          }
+        }
+
+        final int? activeImageMPID =
+            th2FileEditController.stateController.imageOperationImageMPID;
+
+        if (activeImageMPID != null) {
+          final MPRuntimeImageInsertConfigMixin activeImage =
+              th2FileEditController.th2File.imageByMPID(activeImageMPID);
+          final String? overlayLabel = th2FileEditController.stateController
+              .getImageOperationOverlayLabelForImage(activeImageMPID);
+
+          if (overlayLabel != null) {
+            widgets.add(
+              MPImageOperationOverlayWidget(
+                key: ValueKey('image_operation_overlay_$activeImageMPID'),
+                th2FileEditController: th2FileEditController,
+                image: activeImage,
+                label: overlayLabel,
+                previewOffset: th2FileEditController.stateController
+                    .getImageOperationPreviewOffsetForImage(activeImageMPID),
               ),
             );
           }